Move to TypeScript
