+++ src/graphql.ts
@@ -1,17 +1,18 @@
+import { request as Request } from '@octokit/request'
+import { Endpoint } from '@octokit/request/dist-types/types';
 import GraphqlError from './error'
+import { GraphQlQueryResponse } from './types'
 
 const NON_VARIABLE_OPTIONS = ['method', 'baseUrl', 'url', 'headers', 'request', 'query']
 
 export default function graphql (request, query, options) {
-  if (typeof query === 'string') {
-    options = Object.assign({ query }, options)
-  } else {
-    options = query
-  }
+  options = typeof query === 'string'
+    ? options = Object.assign({ query }, options)
+    : options = query
 
   const requestOptions = Object.keys(options).reduce((result, key) => {
     if (NON_VARIABLE_OPTIONS.includes(key)) {
-      result[key] = options[key]
+      result[key] = options![key]
       return result
     }
 
@@ -19,7 +20,7 @@ export default function graphql (request, query, options) {
       result.variables = {}
     }
 
-    result.variables[key] = options[key]
+    result.variables[key] = options![key]
     return result
   }, {})
