+++ src/index.ts
@@ -1,15 +1,15 @@
 import { request } from '@octokit/request'
 import getUserAgent from 'universal-user-agent'
 
-import { version } from './package.json'
+import { version } from '../package.json'
 const userAgent = `octokit-graphql.js/${version} ${getUserAgent()}`
 
-import withDefaults from './lib/with-defaults'
+import withDefaults from './with-defaults'
 
 export default withDefaults(request, {
-  method: 'POST',
-  url: '/graphql',
   headers: {
     'user-agent': userAgent
-  }
+  },
+  method: 'POST',
+  url: '/graphql'
 })
