+++ src/types.ts
@@ -1,0 +1,12 @@
+export interface GraphQlQueryResponse {
+  data: { [ key: string ]: any } | null
+  errors?: [{
+    message: string
+    path: [string]
+    extensions: { [ key: string ]: any }
+    locations: [{
+      line: number,
+      column: number
+    }]
+  }]
+}
